Move DevRant class back to app module
